% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip-lookup.R
\name{qr_ip_lookup}
\alias{qr_ip_lookup}
\title{Get list of Prefixes/ASNs for selected IP}
\usage{
qr_ip_lookup(ip, api_key = qr_api_key())
}
\arguments{
\item{ip}{IP to lookup}

\item{api_key}{your Qrator Radar API key; see \code{\link[=qr_api_key]{qr_api_key()}}.}
}
\value{
list (use \code{\link[=qr_tidy]{qr_tidy()}} to turn the raw API result into a data frame)
}
\description{
Get list of Prefixes/ASNs for selected IP
}
\examples{
qr_tidy(qr_ip_lookup("17.253.144.10"))
}
