% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{qr_tidy}
\alias{qr_tidy}
\title{Turn a Qrator Radar raw API result into a data frame}
\usage{
qr_tidy(res)
}
\arguments{
\item{res}{raw QRator Radar API result from the query functions in
this package.}
}
\value{
data frame
}
\description{
Turn a Qrator Radar raw API result into a data frame
}
\examples{
qr_tidy(qr_ip_lookup("17.253.144.10"))
}
