% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whois.R
\name{qr_whois}
\alias{qr_whois}
\title{Retrieve the raw WHOIS record for a given autonomous system}
\usage{
qr_whois(asn)
}
\arguments{
\item{asn}{autonomous system number; should be just the number but
the function will also work if the input is prefixed with
"AS" or "as"}
}
\value{
character
}
\description{
Retrieve the raw WHOIS record for a given autonomous system
}
\examples{
qr_whois("7015")
}
