% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qradar-package.R
\docType{package}
\name{qradar}
\alias{qradar}
\alias{qradar-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/qradar}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/qradar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
