% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qradar-package.R
\docType{package}
\name{qradar}
\alias{qradar}
\alias{qradar-package}
\title{Gather Autonomous System, IP Address, and Routing Information from Qrator Radar}
\description{
Qrator has an API (\url{https://api.radar.qrator.net/}) that can
be queried for information on autonomous systems, IP addresses, and
various internet routing metadata. Tools are provided to perform these
queries and retrieve resultsets. Note that an account is needed to
generate the free API key which is required for all API calls
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/qradar}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/qradar/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
